requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-smorest
passlib
python-dotenv
gunicorn
psycopg2-binary
flask-cors
flask-migrate
